creating sample_app_donate.md
Hi all.

This file is to demonstrate a conceptual idea of a donation link on apps.fyne.io which was something I had in mind. There is a property in the markdown file called ``donation``, where the user enters the payment link. I am not recommending any payment company, product or service. In that sense, everything is like 'sample_app_donate' which is a general idea.

*This file serves only to demonstrate the donation link. So... I hope to help in some way with this repository.*