update app.html
Hi all.

On line 13 I added the same code as on the previous line, in this case line 12. On line 13 I added a donation link. The code that generates the donation link is the same code that generates the home link for each repository.

In theory, if everything is right, the donation link may or may not appear in each repository. This "problem" is expected, because the developer must specify the donation link individually in the file, otherwise, some markdown files will display the donation link and some will not.

Well I say that, because I initially tested this idea on a server.

I'm waiting for a new feedback if the idea is good or not, and if it was implemented successfully or not.

*Anyway I would be happy to help, and if something is wrong, I apologize.*